[lmj] StyledComponent 분리함 - 알고리즘 연결 부분 미완성이라 오류뜸 (실행에 문제 되는 건 아니라고 판단하여 올림)
