+++ front-end/src/app/components/Mymenu.jsx
@@ -1,6 +1,4 @@
-import React from 'react';
-import { NavLink as BaseNavLink } from 'react-router-dom';
-import { styled } from 'styled-components';
+import { MenuLi, MenuUl, NavStyle } from 'app/style/StyleMypage';
 
 const Mymenu = () => {
   return (
@@ -15,31 +13,4 @@ const Mymenu = () => {
   );
 };
 
-const MenuUl = styled.ul`
-  min-width: 220px;
-  height: 100px;
-  text-align: center;
-`;
-const MenuLi = styled.li`
-  position: relative;
-`;
-const NavStyle = styled(BaseNavLink)`
-  display: block;
-  padding: 6px 15px;
-  border-left: 3px solid;
-  font-size: 1.3em;
-  color: ${({ theme }) => theme.light.t01};
-  font-weight: 500;
-  &:hover,
-  &:focus {
-    color: ${({ theme }) => theme.color.c01};
-  }
-  &:hover::before {
-    background-color: ${({ theme }) => theme.color.c07};
-  }
-  &.active {
-    color: ${({ theme }) => theme.color.c07};
-    font-weight: 700;
-  }
-`;
 export default Mymenu;
