+++ front-end/src/app/pages/Algorithms/Algorithms.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import AlgorithmsMenu from 'app/components/Algorithms/Menu';
+import AlMenu from 'app/components/Algorithms/AlMenu';
 import { styled } from 'styled-components';
 
 const Algorithms = () => {
   return (
     <AlContainer>
-      <AlgorithmsMenu />
+      <AlMenu />
       <Outlet style={{ width: (theme) => theme.common.col9 }} />
     </AlContainer>
   );
@@ -15,7 +15,8 @@ const AlContainer = styled.div`
   display: flex;
   > div:last-child {
     width: ${({ theme }) => theme.common.col9};
-    padding: 4.4rem 1.3rem 0 3.3rem;
+    // padding: 4.4rem 1.3rem 0 3.3rem;
+    padding: 0 1.3rem 0 3.3rem;
   }
 `;
 export default Algorithms;
