+++ front-end/src/app/pages/Algorithms/Child.jsx
@@ -1,17 +1,11 @@
 // 알고리즘 - 문제추천 - 알고리즘
 import { Autocomplete, FormControl } from '@mui/joy';
 import RankLevel from 'app/components/Algorithms/RankLevel';
-import { Table } from 'app/components/Board/Table';
-import { SubmitButton } from 'app/style/StyledComponent';
+import { Table } from 'app/components/Table';
+import { ResetButton, Top } from 'app/style/StyleAlgorithms';
 import Cookies from 'js-cookie';
 import { styled } from 'styled-components';
 
-const StepTop = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-bottom: 0.4rem;
-`;
 const Right = styled.div`
   display: flex;
   height: 40px;
@@ -20,12 +14,9 @@ const AutocompleteStyle = styled(Autocomplete)`
   width: 100%;
   border-color:  ${({ theme }) => theme.light.t03}; !important;
 `;
-const ResetButton = styled(SubmitButton)`
+const Button = styled(ResetButton)`
   width: ${({ theme }) => theme.common.col3};
   padding: 6px 10px;
-  margin: 0;
-  border-radius: 8px;
-  font-size: ${({ theme }) => theme.common.base};
 `;
 
 const Child = () => {
@@ -49,7 +40,7 @@ const Child = () => {
 
   return (
     <div>
-      <StepTop>
+      <Top>
         <RankLevel level={level}></RankLevel>
         <Right>
           <AutocompleteStyle
@@ -63,9 +54,9 @@ const Child = () => {
           >
             <FormControl id="multiple-limit-tags"></FormControl>
           </AutocompleteStyle>
-          <ResetButton>새로고침</ResetButton>
+          <Button>새로고침</Button>
         </Right>
-      </StepTop>
+      </Top>
       {Table(['문제 번호', '제목', '해결 여부'], modifiedDataList, 'https://www.acmicpc.net/board/view', pageNumber)}
     </div>
   );
