+++ front-end/src/app/pages/Algorithms/Rank.jsx
@@ -1,16 +1,18 @@
 // 알고리즘 - 랭킹분석
-import { styled } from 'styled-components';
+import { setMemberName } from 'app/store';
+import { AlgorithmsRank, Img, LinkStyle, Rating, RatingBefore, RatingContent, RatingLi, TypeH2, Value } from 'app/style/StyleAlgorithms';
 
 const Rank = () => {
+  const userName = setMemberName();
   return (
-    <AlgorithmsType>
+    <AlgorithmsRank>
       <div style={{ marginBottom: '6rem' }}>
-        <H2>기간별 진행도</H2>
-        <Img src="https://ghchart.rshah.org/219138/OlMinje" alt="백준 잔디" />
+        <TypeH2>기간별 진행도</TypeH2>
+        <Img src={`https://ghchart.rshah.org/219138/${userName}`} alt="백준 잔디" />
         <LinkStyle onClick={() => (document.location.href = 'https://github.com/Bulldozo')}>깃허브 바로가기</LinkStyle>
       </div>
       <div style={{ width: '40%', margin: '0 auto' }}>
-        <H2>Rating</H2>
+        <TypeH2>Rating</TypeH2>
         <Rating>
           <RatingBefore>Rank</RatingBefore>
           <RatingContent style={{ borderBottom: '1px solid #ccc' }}>
@@ -31,76 +33,8 @@ const Rank = () => {
           </RatingContent>
         </Rating>
       </div>
-    </AlgorithmsType>
+    </AlgorithmsRank>
   );
 };
-const AlgorithmsType = styled.div`
-  width: auto;
-  min-height: 600px;
-  margin: 0 auto;
-  padding: 120px 0 0;
-  text-align: center;
-`;
-const H2 = styled.h2`
-  font-size: 40px;
-  line-height: 50px;
-  font-weight: 500;
-  text-align: center;
-  margin-bottom: 20px;
-`;
-const Img = styled.img`
-  display: block;
-  width: 80%;
-  height: 150px;
-  margin: 0 auto 1rem;
-`;
-const LinkStyle = styled.a`
-  display: inline-block;
-  background: none;
-  text-decoration: underline;
-  font-size: 15px;
-  color: #1976d2;
-  cursor: pointer;
-`;
-const Rating = styled.div`
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  box-shadow: 0 0 0 2px ${({ theme }) => theme.color.c01};
-  border-radius: 10px;
-  height: 200px;
-`;
-const RatingBefore = styled.div`
-  position: absolute;
-  top: calc(50% - 30px);
-  left: calc(50% - 30px);
-  width: 60px;
-  height: 60px;
-  padding: 19px 0;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.color.c05};
-  font-weight: 500;
-  color: ${({ theme }) => theme.light.b03};
-`;
-const RatingContent = styled.ul`
-  height: 50%;
-  display: flex;
-  box-sizing: border-box;
-  align-items: center;
-  text-align: left;
-`;
-const RatingLi = styled.li`
-  width: 50%;
-  height: 100%;
-  padding: 40px 25px;
-  font-weight: 700;
-`;
-const Value = styled.span`
-  float: right;
-  font-size: 25px;
-  line-height: 25px;
-  font-weight: 500;
-  color: ${({ theme }) => theme.color.c08};
-`;
+
 export default Rank;
