+++ front-end/src/app/pages/Algorithms/Step.jsx
@@ -1,22 +1,12 @@
 // 알고리즘 - 문제추천 - 실력별
 import RankLevel from 'app/components/Algorithms/RankLevel';
-import { Table } from 'app/components/Board/Table';
-import { SubmitButton } from 'app/style/StyledComponent';
+import { Table } from 'app/components/Table';
+import { ResetButton, Top } from 'app/style/StyleAlgorithms';
 import Cookies from 'js-cookie';
 import { styled } from 'styled-components';
 
-const StepTop = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-end;
-  margin-bottom: 0.4rem;
+const StepTop = styled(Top)`
   height: 85px;
-`;
-const ResetButton = styled(SubmitButton)`
-  padding: 6px 15px;
-  margin: 0;
-  border-radius: 8px;
-  font-size: ${({ theme }) => theme.common.base};
 `;
 
 const Step = () => {
