+++ front-end/src/app/pages/Algorithms/Type.jsx
@@ -17,6 +17,7 @@ const Wrap = styled.div`
   width: 100%;
   height: 500px;
 `;
+
 const exData = [
   {
     taste: 'fruity',
