+++ front-end/src/app/pages/Algorithms/Wrong.jsx
@@ -17,6 +17,7 @@ const Wrap = styled.div`
   width: 100%;
   height: 500px;
 `;
+
 const exData = [
   {
     id: 'python',
