+++ front-end/src/app/pages/Board/BoardList.jsx
@@ -1,8 +1,8 @@
-import { Table } from 'app/components/Board/Table';
+import { Table } from 'app/components/Table';
 import instance from 'app/slices/Instance';
 import { useCallback, useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { styled } from 'styled-components';
+import { Board, Wrap, StyleP, Tab, StyledUl, StyleLi, StyledButton, Pagination, PaginButton } from 'app/style/StyleBoard';
 
 const BoardList = () => {
   const navigate = useNavigate();
@@ -90,80 +90,4 @@ const BoardList = () => {
   );
 };
 
-const Board = styled.div`
-  max-width: ${({ theme }) => theme.common.col9};
-  margin: 0 auto;
-`;
-const Wrap = styled.div`
-  margin-bottom: 3rem;
-  box-shadow: 0px 10px 2px 5px ${({ theme }) => theme.color.c01};
-`;
-const StyleP = styled.p`
-  padding: 10px 5px;
-  text-align: right;
-  font-size: 1.2rem;
-  font-weight: 400;
-  span {
-    font-size: 1.4rem;
-    font-weight: 600;
-    color: #ff73a5;
-  }
-`;
-const Tab = styled.button`
-  padding: 15px 10px;
-  font-size: 1.1rem;
-  font-weight: 700;
-  width: 100%;
-  &:hover {
-    color: ${({ theme }) => theme.color.c08};
-  }
-  &:focus {
-    color: ${({ theme }) => theme.color.c07};
-    background-color: ${({ theme }) => theme.light.t03};
-  }
-  &.active {
-    color: ${({ theme }) => theme.color.c07};
-    background-color: ${({ theme }) => theme.light.t03};
-  }
-`;
-const StyledUl = styled.ul`
-  display: flex;
-  justify-content: space-between;
-  border-bottom: 1px solid ${({ theme }) => theme.color.c01};
-`;
-const StyleLi = styled.li`
-  width: 33.33%;
-  background-color: ${({ theme }) => theme.light.b03};
-`;
-const StyledButton = styled.button`
-  display: block;
-  margin-left: auto;
-  padding: 0.7rem 1.5rem;
-  border-radius: 10px;
-  background-color: ${({ theme }) => theme.color.c05};
-  font-size: 1rem;
-  font-weight: 500;
-  &:hover {
-    color: ${({ theme }) => theme.light.b01};
-    background-color: ${({ theme }) => theme.light.t02};
-  }
-`;
-const Pagination = styled.div`
-  margin: 0 auto;
-  display: flex;
-  justify-content: center;
-`;
-const PaginButton = styled.button`
-  padding: 0.7rem;
-  margin: 0 5px;
-  border-radius: 5px;
-  background-color: ${({ theme }) => theme.light.b02};
-  font-size: 1rem;
-  font-weight: 500;
-  &:hover,
-  &.active {
-    color: ${({ theme }) => theme.light.b01};
-    background-color: ${({ theme }) => theme.light.t02};
-  }
-`;
 export default BoardList;
