+++ front-end/src/app/pages/auth/ForgotPassword.jsx
@@ -1,6 +1,5 @@
-import { SubmitButton, Title } from 'app/style/StyledComponent';
+import { Button, ForgotContainer, Form, Input, NavBack, StyleTitle } from 'app/style/StyleAuth';
 import { useState } from 'react';
-import { NavLink } from 'react-router-dom';
 import { styled } from 'styled-components';
 
 const ForgotPassword = () => {
@@ -10,12 +9,11 @@ const ForgotPassword = () => {
   const handleSubmit = (e) => {
     e.preventDefault();
     // 비밀번호 찾기 로직 추가 필요
-
     setIsSubmitted(true);
   };
 
   return (
-    <Container>
+    <ForgotContainer>
       {isSubmitted ? (
         <PStyle>비밀번호 재설정 링크가 이메일로 전송되었습니다.</PStyle>
       ) : (
@@ -26,56 +24,26 @@ const ForgotPassword = () => {
             <br />
             비밀번호 재설정 링크를 보내드립니다.
           </PStyle>
-          <Input type="email" id="memberEmail" value={email} onChange={(e) => setEmail(e.target.value)} />
-          <Button type="submit">비밀번호 재설정</Button>
+          <ForgotInput type="email" id="memberEmail" value={email} onChange={(e) => setEmail(e.target.value)} />
+          <StyleButton type="submit">비밀번호 재설정</StyleButton>
         </Form>
       )}
-      <NavStyle to="/members/login">돌아가기</NavStyle>
-    </Container>
+      <NavBack to="/members/login">돌아가기</NavBack>
+    </ForgotContainer>
   );
 };
-const Container = styled.div`
-  padding: 100px 10px;
-  text-align: center;
-  * {
-    font-weight: 500;
-  }
-`;
-const StyleTitle = styled(Title)`
-  margin-bottom: 2rem;
-  color: ${({ theme }) => theme.light.t01};
-  text-align: center;
-`;
-const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-const Input = styled.input`
+
+const ForgotInput = styled(Input)`
   width: calc(50% - 200px);
-  height: 48px;
-  padding: 0 10px;
   margin-bottom: 1rem;
-  border-radius: 6px;
-  box-sizing: border-box;
-  background-color: ${({ theme }) => theme.light.t02};
-  ::placeholder {
-    color: ${({ theme }) => theme.color.c01};
-  }
 `;
 const PStyle = styled.p`
   margin-bottom: 2rem;
   font-size: 1.1rem;
   line-height: 1.4rem;
 `;
-const Button = styled(SubmitButton)`
+const StyleButton = styled(Button)`
   width: calc(50% - 200px);
   margin-bottom: 1rem !important;
-  border-radius: 6px !important;
-  color: ${({ theme }) => theme.light.t03} !important;
-`;
-const NavStyle = styled(NavLink)`
-  text-decoration: underline;
-  color: ${({ theme }) => theme.color.c05};
 `;
 export default ForgotPassword;
