+++ front-end/src/app/pages/auth/MyCertify.jsx
@@ -1,26 +1,49 @@
-// 백준 연동
-// [ ]  구조 바꿔야함
-// import { useSelector } from 'react-redux';
 import logo from 'app/assets/images/logo.svg';
 import Mymenu from 'app/components/Mymenu';
-import { SubmitButton, Title } from 'app/style/StyledComponent';
-import { styled } from 'styled-components';
+import { getBackUsers } from 'app/slices/BackUserSlice';
+import { setMemberName } from 'app/store';
+import { SubmitButton } from 'app/style/GlobalStyle';
+import {
+  After,
+  Before,
+  BlueText,
+  Circle,
+  Container,
+  Content,
+  Img,
+  PinkText01,
+  PinkText02,
+  PinkText03,
+  Right,
+  StyleLi,
+  StyleP,
+  StyleTitle,
+  StyleUl,
+  WhiteText,
+} from 'app/style/StyleMypage';
+import { useEffect, useState } from 'react';
+import { useDispatch } from 'react-redux';
 
 const MyCertify = () => {
-  // [ ] 작업 예정 코드 다 삭제할거임
-  // const user = useSelector((state) => state.user);
-  // console.log('User:{}', user);
+  const dispatch = useDispatch();
+  const [usersName, setUsersName] = useState(null);
+  // const [info, setInfo] = useState([]);
 
-  const test = true;
+  useEffect(() => {
+    setUsersName(setMemberName());
+    if (usersName !== null) {
+      dispatch(getBackUsers(usersName));
+    }
+  }, [dispatch, usersName]);
 
-  if (test) {
-    // 연동 후
-    return (
-      <Container>
-        <StyleTitle>내정보</StyleTitle>
-        <Content>
-          <Mymenu />
-          <Right>
+  // 연동 후
+  return (
+    <Container>
+      <StyleTitle>내정보</StyleTitle>
+      <Content>
+        <Mymenu />
+        <Right>
+          {usersName !== null ? (
             <After>
               <div style={{ width: '30%' }}>
                 <Circle></Circle>
@@ -45,145 +68,26 @@ const MyCertify = () => {
                 </div>
               </div>
             </After>
-          </Right>
-        </Content>
-      </Container>
-    );
-  } else {
-    // 연동 전
-    return (
-      <Container>
-        <StyleTitle>내정보</StyleTitle>
-        <Content>
-          <Mymenu />
-          <Right>
+          ) : (
             <Before>
               <Img src={logo} alt="logo" />
-              <PStyle>
+              <StyleP>
                 해당 페이지는 백준 아이디를 연동해야만 사용 가능합니다.
                 <br />
                 연동에 성공하면 해당 기능을 사용할 수 있습니다.
-              </PStyle>
-              <UlStyle>
-                <LiStyle>사용 가능 기능1</LiStyle>
-                <LiStyle>사용 가능 기능2</LiStyle>
-                <LiStyle>사용 가능 기능3</LiStyle>
-              </UlStyle>
+              </StyleP>
+              <StyleUl>
+                <StyleLi>사용 가능 기능1</StyleLi>
+                <StyleLi>사용 가능 기능2</StyleLi>
+                <StyleLi>사용 가능 기능3</StyleLi>
+              </StyleUl>
               <SubmitButton style={{ color: 'var(--text-100)', width: '40%' }}>백준 연동하러 가기</SubmitButton>
             </Before>
-          </Right>
-        </Content>
-      </Container>
-    );
-  }
-};
-const Container = styled.div`
-  padding: 80px 0 0;
-`;
-const StyleTitle = styled(Title)`
-  width: calc(100% - 400px);
-  margin: 0 auto 1rem;
-  text-align: left;
-  font-size: 45px;
-  line-height: 55px;
-  font-weight: 500;
-`;
-const Content = styled.div`
-  display: flex;
-  width: calc(100% - 400px);
-  margin: 0 auto 2rem;
-`;
-const Right = styled.div`
-  width: 100%;
-  min-height: 600px;
-`;
-const Before = styled.div`
-  display: none;
-  height: 100%;
-  padding: 100px 10px;
-  border: 1px solid ${({ theme }) => theme.color.c01};
-  border-radius: 10px;
-  text-align: center;
-`;
-const After = styled(Before)`
-  display: flex;
-  width: 90%;
-  height: 70%;
-  margin: 0 auto;
-  padding: 40px;
-  background-image: linear-gradient(
-    130deg,
-    ${({ theme }) => theme.light.t03},
-    ${({ theme }) => theme.color.c03},
-    ${({ theme }) => theme.light.t03}
+          )}
+        </Right>
+      </Content>
+    </Container>
   );
-`;
-const Img = styled.img`
-  width: 150px;
-  padding: 10px;
-  vertical-align: middle;
-  border-radius: 10px;
-  background-color: ${({ theme }) => theme.light.t03};
-`;
-const PStyle = styled.p`
-  margin: 4rem 0 1rem;
-  font-size: 1.2rem;
-  font-weight: 500;
-`;
-const UlStyle = styled.ul`
-  counter-reset: number 0;
-  margin-bottom: 4rem;
-`;
-const LiStyle = styled.li`
-  position: relative;
-  &::before {
-    position: absolute;
-    top: 50%;
-    left: calc(50% - 65px);
-    transform: translateY(-50%);
-    counter-increment: number 1;
-    content: counter(number) '.';
-    color: 
-    ${({ theme }) => theme.light.t01}
-    font-size: 0.8em;
-    z-index: 1;
-  }
-`;
-const Circle = styled.p`
-  width: 200px;
-  height: 200px;
-  margin: 0 auto 2rem;
-  border-radius: 50%;
-  background-color: ${({ theme }) => theme.light.b01};
-`;
-const PinkText01 = styled.p`
-  width: 100%;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  overflow: hidden;
-  font-weight: 500;
-  font-size: 25px;
-  color: ${({ theme }) => theme.color.c07};
-`;
-const PinkText02 = styled(PinkText01)`
-  margin-bottom: 11.5rem;
-  font-size: 30px;
-`;
-const PinkText03 = styled(PinkText01)`
-  display: inline;
-  font-size: 20px;
-`;
-const BlueText = styled.span`
-  margin-bottom: 11.5rem;
-  font-size: 2rem;
-  font-weight: 700;
-  color: ${({ theme }) => theme.color.c04};
-`;
-const WhiteText = styled.p`
-  display: inline-block;
-  width: 40%;
-  font-weight: 500;
-  font-size: 20px;
-  color: ${({ theme }) => theme.light.b01};
-`;
+};
+
 export default MyCertify;
