--- front-end/src/app/style/ThemeProvider.js
@@ -1,33 +1,0 @@
-// [ ] 다크 모드 코드 미리 세팅 - 테스트 아직 안 함
-
-import { lightTheme, darkTheme } from '../theme/theme';
-import { createContext, useState, useContext, useCallback } from 'react';
-import { ThemeProvider as StyledProvider } from 'styled-components';
-
-export const ThemeProvider = ({ children }) => {
-  const [ThemeMode, setThemeMode] = useState('light');
-  const themeObject = ThemeMode === 'light' ? lightTheme : darkTheme;
-
-  return (
-    <ThemeContext.Provider value={{ ThemeMode, setThemeMode }}>
-      <StyledProvider theme={themeObject}>{children}</StyledProvider>
-    </ThemeContext.Provider>
-  );
-};
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  const { ThemeMode, setThemeMode } = context;
-
-  const toggleTheme = useCallback(() => {
-    if (ThemeMode === 'light') {
-      setThemeMode('dark');
-    } else {
-      setThemeMode('light');
-    }
-  }, [ThemeMode]);
-
-  return [ThemeMode, toggleTheme];
-};
-
-export { ThemeProvider, useTheme };
